Extract utilities from plotting module.
